extract SubmissionModel
