+++ app/Models/ProblemModel.php
@@ -5,6 +5,7 @@ namespace App\Models;
 use GrahamCampbell\Markdown\Facades\Markdown;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\DB;
+use App\Models\Submission\SubmissionModel;
 use Cache;
 
 class ProblemModel extends Model
